# generated/openapi/knowledge-fabric.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenHuman Knowledge Fabric API",
    "version": "1.1.0-draft.1",
    "description": "Generated from ontology source digest e2e0283906bed576d89acee4e409cb14e475f04d4aad94bd80178f3f26b5afb9."
  },
  "paths": {
...
    "/actions/attach_evidence": {
      "post": {
        "summary": "Execute attach_evidence",
        "operationId": "attach_evidence",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Action"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Applied, or replayed from the idempotency key"
          },
          "403": {
            "description": "Actor lacks the authority for this action"
          },
          "409": {
            "description": "Precondition failed — state or version moved"
          },
          "422": {
            "description": "Input failed schema or invariant validation"
          }
        }
      }
    },
...
    "/actions/add_controlled_document": {
      "post": {
        "summary": "Execute add_controlled_document",
        "operationId": "add_controlled_document",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Action"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Applied, or replayed from the idempotency key"
          },
          "403": {
            "description": "Actor lacks the authority for this action"
          },
          "409": {
            "description": "Precondition failed — state or version moved"
          },
          "422": {
            "description": "Input failed schema or invariant validation"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvidenceReference": {
        "type": "object",
        "properties": {
          "node_id": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "uri": {
            "type": "string",
            "format": "uri"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "media_type": {
            "type": "string"
          }
        },
        "minProperties": 1,
        "additionalProperties": false
      },
...
      "Action": {
        "type": "object",
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.1.0-draft.1"
          },
          "action_id": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "action_type": {
            "type": "string",
            "enum": [
              "create_initiative",
              "triage_initiative",
              "authorize_project",
              "activate_project",
              "create_work_package",
              "start_work_package",
              "accept_work_package",
              "issue_work_order",
              "accept_work_order",
              "amend_work_order",
              "submit_work_execution",
              "review_work_execution",
              "issue_acceptance",
              "propose_decision",
              "accept_decision",
              "reject_decision",
              "supersede_decision",
              "open_change",
              "approve_change",
              "verify_change",
              "make_change_effective",
              "submit_invoice",
              "approve_invoice",
              "authorize_payment",
              "record_payment_settlement",
              "reconcile_payment",
              "complete_project_technical",
              "close_project_administrative",
              "attach_evidence",
              "correct_record",
              "promote_configuration_item",
              "supersede_configuration_item",
              "retire_configuration_item",
              "publish_interface_contract",
              "deprecate_interface_contract",
              "withdraw_interface_contract",
              "record_physical_binding",
              "remove_physical_binding",
              "add_controlled_document",
              "submit_document_for_review",
              "approve_controlled_document",
              "make_document_effective",
              "supersede_controlled_document",
              "withdraw_controlled_document",
              "raise_nonconformity",
              "contain_nonconformity",
              "investigate_nonconformity",
              "disposition_nonconformity",
              "close_nonconformity",
              "open_capa",
              "approve_capa_plan",
              "implement_capa",
              "check_capa_effectiveness",
              "close_capa",
              "register_supplier",
              "qualify_supplier",
              "restrict_supplier",
              "disqualify_supplier",
              "register_equipment",
              "place_equipment_in_service",
              "remove_equipment_from_service",
              "quarantine_equipment",
              "retire_equipment",
              "receive_complaint",
              "triage_complaint",
              "investigate_complaint",
              "close_complaint",
              "propose_risk_control",
              "implement_risk_control",
              "verify_risk_control",
              "retire_risk_control",
              "define_test",
              "approve_test_definition",
              "supersede_test_definition",
              "plan_test_execution",
              "execute_test",
              "record_test_result",
              "invalidate_test_execution"
            ]
          },
          "actor": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "acting_roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "uniqueItems": true
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time"
          },
          "transaction_id": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "idempotency_key": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "targets": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            },
            "minItems": 1,
            "uniqueItems": true
          },
          "input": {
            "type": "object"
          },
          "preconditions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "result": {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "applied",
                  "rejected",
                  "failed"
                ]
              },
              "message": {
                "type": "string"
              },
              "created_nodes": {
                "type": "array",
                "items": {
                  "type": "string",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                }
              },
              "created_edges": {
                "type": "array",
                "items": {
                  "type": "string",
                  "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
                }
              },
              "before_digest": {
                "type": "string",
                "pattern": "^[0-9a-f]{64}$"
              },
              "after_digest": {
                "type": "string",
                "pattern": "^[0-9a-f]{64}$"
              }
            },
            "required": [
              "status"
            ],
            "additionalProperties": false
          },
          "reason": {
            "type": "string"
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidenceReference"
            }
          }
        },
        "required": [
          "schema_version",
          "action_id",
          "action_type",
          "actor",
          "acting_roles",
          "occurred_at",
          "transaction_id",
          "idempotency_key",
          "targets",
          "input",
          "result"
        ],
        "additionalProperties": false
      },
...
    }
  }
}